django/myproject/requirements.txt
Django
requests
django-cors-headers
djangorestframework
django-filter
celery
python-memcached
celery==5.2.7 
kombu==5.2.4 
redis==4.5.5
psycopg2-binary